Add dashboard route and redirect login flow
