+++ src/lib/auth.ts
@@ -1,0 +1,1 @@
+export const AUTH_TOKEN_STORAGE_KEY = "morelibre.auth_token";
